test(nutrition): consolidate PRO targets coverage
Keep the full production-path branch matrix in the CI-selected edge suite while preserving the canonical ownership and exact-alias architecture guards. This removes the redundant standalone service test file so the PR stays within the standard 20-file scope cap without changing runtime behavior.
